Updated to work with redmine 2.x
Updated to work with redmine 2.x / rails 3.x only.  Changes between rails 2 and 3 make it difficult to support both versions at the same time.

NOTE: In order for this plugin to work correctly, you must comment out the
default root and login routes in the redmine app's config/routes.db.  This
is due to the fact that those routes are matched before the plugin's routes.

TODO: Find a way to override the default routes from the redmine app so that we do not need to make any modifications to the redmine application directly.